src/test/kotlin/Day21Test.kt
import org.eruanno.Day21
import org.junit.jupiter.api.Assertions.assertEquals
import org.junit.jupiter.params.ParameterizedTest
import org.junit.jupiter.params.provider.CsvSource

class Day21Test {
    private val subject = Day21()

    @ParameterizedTest
    @CsvSource(
        "'day-21-test', 126384",
        //"'day-21', 0"
    )
    fun testPartOne(testFile: String, expectedResult: Int) {
        assertEquals(expectedResult, subject.partOne(testFile))
    }

    @ParameterizedTest
    @CsvSource(
        "'day-21-test', '0'",
        "'day-21', '0'"
    )
    fun testPartTwo(testFile: String, expectedResult: Int) {
        assertEquals(expectedResult, subject.partTwo(testFile))
    }
}
